add simple argparse and init graphic
