+++ README.md
@@ -1,4 +1,14 @@
 # FRA-P-RS (France-parse-addresses)
+ _______ .______          ___           .______      ___      .______          _______.
+|   ____||   _  \        /   \          |   _  \    /   \     |   _  \        /       |
+|  |__   |  |_)  |      /  ^  \   ______|  |_)  |  /  ^  \    |  |_)  |      |   (----`
+|   __|  |      /      /  /_\  \ |______|   ___/  /  /_\  \   |      /        \   \    
+|  |     |  |\  \----./  _____  \       |  |     /  _____  \  |  |\  \----.----)   |   
+|__|     | _| `._____/__/     \__\      | _|    /__/     \__\ | _| `._____|_______/    
+                                                                                       
+Fra-parse 0.2.0
+Developed by eCecchetti
+
 Simple script that permits to parse addresses in France format.
 
 ## Install
@@ -50,12 +60,51 @@ Details: {'urba_names': ['rue'], 'prepositions': ['de'], 'city': ['colombes'], '
 Unparsed string remained is: 
 Parse quality scored: 1.0
 ```
+The script runned with `poetry run script` can execute in 3 different ways:
+
+### Single address
+with the command:
+`poetry run script -a "Rue, Saint-Philippe 31 (De)"`
+```text
+Fra-parse 0.2.0
+Developed by eCecchetti
+Initial address is: Rue, Saint-Philippe 31 (De)
+Result is: 31 Rue De Saint Philippe 
+Details: {'urba_names': ['rue'], 'prepositions': ['de'], 'city': [], 'addres_num': ['31'], 'department': [], 'street_name': ['saint philippe']}
+Unparsed string remained is: ,
+Parse quality scored: 0.98
+Process has finished!
+```
+
+### List of addresses
+with the command:
+`poetry run script -l "Rue, Saint-Philippe 31 (De); Saint Marcelle (De) Avenue 32 10141"`
+```text
+Fra-parse 0.2.0
+Developed by eCecchetti
+Processing: 100%|████████████████████████████████████████████████| 2/2 [00:00<00:00, 134.66item/s]
+Unparsed: Rue, Saint-Philippe 31 (De) --> Parsed: 31 Rue De Saint Philippe
+Unparsed:  Saint Marcelle (De) Avenue 32 10141 --> Parsed: 32 Avenue De Saint Marcelle 10141
+Process has finished!
+```
+
+### File conversion
+with the command:
+`poetry run script -i "Rue, Saint-Philippe 31 (De); Saint Marcelle (De) Avenue 32 10141"`
+Fra-parse 0.2.0
+Developed by eCecchetti
+Found 234253 addresses to parse
+Processing: 100%|████████████████████████████████████████████████| 234253/234253 [27:18<00:00, 142.99item/s]
+Storing into file the results
+Result printed at:  frapars/res/parsed_addresses.csv
+Process has finished!
+
 
 ## To-Do List
 - [x] Publish in git
 - [x] Clean structure (use maybe poetry)
-- [ ] Create a simple init graphic
-- [ ] Add args and helper: [`single_address`, `csv`]
+- [x] Create a simple init graphic
+- [x] Add args and helper: [`single_address`, `list_of_addresses`, `csv`]
 - [ ] Split the csv in piece and execute in more thread  (to fasten up)
 - [ ] Write junit tests
 - [ ] Create package for pypi to be imported
