+++ banner.txt
@@ -1,0 +1,9 @@
+ _______ .______          ___           .______      ___      .______          _______.
+|   ____||   _  \        /   \          |   _  \    /   \     |   _  \        /       |
+|  |__   |  |_)  |      /  ^  \   ______|  |_)  |  /  ^  \    |  |_)  |      |   (----`
+|   __|  |      /      /  /_\  \ |______|   ___/  /  /_\  \   |      /        \   \    
+|  |     |  |\  \----./  _____  \       |  |     /  _____  \  |  |\  \----.----)   |   
+|__|     | _| `._____/__/     \__\      | _|    /__/     \__\ | _| `._____|_______/    
+                                                                                       
+Fra-parse ${application.version}
+Developed by eCecchetti
