--- frapars/__init__.py
@@ -1,5 +1,0 @@
-"""
-The fra parser version
-"""
-
-__version__ = '0.1.0'
